+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
-from model.grid import Grid, GridImplWithPythonList
+from model.grid import Grid
+from model.grid_impl_with_python_list import GridImplWithPythonList
 from model.utils import Point2D
